+++ src/main/kotlin/com/mkroo/lmsdemo/dto/LectureOpeningRequest.kt
@@ -2,6 +2,6 @@ package com.mkroo.lmsdemo.dto
 
 data class LectureOpeningRequest(
     val title: String,
-    val maxStudentCount: Long,
+    val maxStudentCount: Int,
     val price: Long
 )
